fix!: detect case sensitive filenames to validate
Fix #207

Now `frecency.validate_database` detects case sensitive filenames and
delete if it is not found. For example, now it deletes `foo.lua` even if
`Foo.lua` still exists.

For performance issue, this function is now an async function, so you
should wrap it with `async.void` to call.
